Move public static methods out of FunctionRegistry
As a preparation step for removing references to FunctionRegistry outside
of the package, move public static methods in FunctionRegistry to
Operatorsignatureutil and LiteralEncoder.
